requirements.txt
aiogram==3.13.0
peewee==3.17.6
python-dotenv==1.0.1
fastapi==0.111.0
httpx==1.0.0b0
